# LocalGems.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LocalGems.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/JobOffers": {
      "get": {
        "tags": [
          "JobOffers"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOffer"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOffer"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOffer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "JobOffers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobOffer"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JobOffer"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/JobOffer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobOffers/{id}": {
      "get": {
        "tags": [
          "JobOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "JobOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobOffer"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JobOffer"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/JobOffer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "JobOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JobOffer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}